Fix an issue with clusters of hardclipped reads
Clusters of hardclipped reads didn't get a five prime or three prime
position assigned, breaking the GFF output.
This was caused by the following logic:
To identify three prime or five prime breakpoints, we scan for reads
with a softclip that starts or ends their alignment at the breakpoint.
To not count the same read for a fiveprime and threeprime break,
we require the aligned portion of the read to not start at 0 (inferred
by the query_alignment_start) if it is counting for the 5 prime break,
and if it is counting for the 3 prime break is should not end at the end
of the length of the aligned read.
Hardclipped reads however do always
start their aligned portion at 0 end end at the length of the read. But
we can count them if they have a left or right hardclip, so that's what
we do now.

A new testcase is included. To facilitate testing there is a new
MockAlignedSegemnt class in test.helpers and all instances of qstart or
qend have been changed to query_alignment_start or query_alignment_end
respecively.