Keep suppl. reads without primary reads by default
We may legitimately be interested in these reads, e.g if the primary
read pair is a TE aligning read and a proper pair. In that case we would
discard the primary read even if it is tagged, to avoid annotating minor
differences inside annotated TEs. The supplementary alignment however
may point to a true TE insertion event, so we should keep it.