Add a genotyping module
that is based on http://dx.doi.org/10.1186/1471-2164-15-795.
The evidence is collected using fetch on indexed BAM files, so
additional changes were necessary. If BamAlignmentReader is called with
`with_index=True` a pysam.AlignmentFile will be returned, since `fetch`
does not work with SAM files.
Additionally accept mate tags if a pair is not a proper pair.