Refine cluster splitting
Previously a cluster of TE reads would be split if the orientation of
reads would switch from Reverse to Forward (which should not happen if
there is no closeby TE and no sequencing artefact caused wrong
orientation of reads). To avoid wrongly splitting clusters due to a
stray read all reads that contribute to the cluster are assembled into
contigs. Reads that may contribute to a new cluster (cluster_b in the
code) will be kept in the original cluster if the align to one of the
contigs that is clearly part of cluster_a.