Add pysamtools_view command
This command can be used to extract test-data from a specific region.
The output bam file contains this region exclusively, and the
coordinates are adjusted so that the start of the specified region
starts at coordinate 0 on the first chromosome. Paired with samtools
faidx one can easily construct a reference genome for this BAM file.