Tune clusterfinding for misaligned long reads
In regions of low complexity repeats (ATATAT, rover's favorite integration
site) a subset of nanopore reads is not clipped at the exact right site. We
allow some wigglespace (10 nt) now for assigning a read to a specific TSD side,
if it is clear that the read belongs to a specific side of the insertion (more
than 100 nt aligned on that side).