Add read to cluster if AD tag makes sense
AD tags may apply only to subset of possible alignments of a read. We only add
an aligned segment with an AD tag to a cluster if the AD tag makes sense. In
practice that means the cigar portion of the tag describes a portion of the
read that is not aligned.