Revert "Try replacing cap3 scaffolding with IDBA"
This reverts commit 8a331347ec078cb7548283c0f4ecbd66eca6a8f4.

IDBA often crashes or produces no contigs, cap3 seems to be better
at this time. Spades works well but is really slow (1.7 seconds for ~40 reads).
Given ~100000 potential sites this would be a bottleneck.
I'll probably try to implement an assembler/scaffolder myself given the very
specifc task at hand. I know the start, and I know that the coverage will drop
at the end, and I know that the orientation of the mate sequences inverts.
With those constraints even a python "assembler" should be able to resolve this
quickly.