Split chunk if we are beyoing target read.
Should solve issues if annotating file a with file b, where certain reads
are missing in file a (e.g markdup, or filtering out unaligned reads).
We know that the input must be queryname sorted, so we can just make
sure that the current read wouldn't sort beyond the target read.